hitchi/pyproject.toml
[tool.poetry]
name = "hitchi"
version = "0.1.0"
description = ""
authors = ["Jonas Groeger <jonas@huntun.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyserial = "^3.5"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
